requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
dj-database-url==0.4.2
Django==1.11.5
djangorestframework==3.6.4
gunicorn==19.7.1
httpie==0.9.9
httpie-django-auth==0.1.1
idna==2.6
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
urllib3==1.22
whitenoise==3.3.1
psycopg2==2.6.2
